fix: more improvements & docs
